+++ inst/shiny_app/server/input_settings.R
@@ -1,4 +1,7 @@
+# Input settings server function -----------------------------------------------
 input_settings <- function(input, output, session) {
+
+    # Uploaded files meta information ------------------------------------------
     files_meta <- reactive({
         res <- list()
         file_input_meta <- input[["file_upload"]]
@@ -10,6 +13,8 @@ input_settings <- function(input, output, session) {
             file_path <- single_file_input_meta[["datapath"]]
 
             single_res <- list(
+                "input_id" = sprintf("is_row%d", i),
+                "file_name" = file_name,
                 "is_ok" = NULL,
                 "error_messages" = NULL,
                 "sequence_length" = NULL,
@@ -17,7 +22,7 @@ input_settings <- function(input, output, session) {
             )
 
             single_file_data <- fread(file_path) # TODO: tryCatch this.
-            single_res[1:2] <- verify_iao_data(single_file_data)
+            single_res[c("is_ok", "error_messages")] <- verify_iao_data(single_file_data)
             if (single_res[["is_ok"]]) {
                # sequence_length
                # protein_state_mapping
@@ -29,9 +34,30 @@ input_settings <- function(input, output, session) {
         res
     })    
 
-    observe(print(files_meta())) # TODO: remove after finishing development.
 
-    # TODO: single row generation as a separate function inside the package (?).
+    # Import summary table UI --------------------------------------------------
+    output[["input_summary_table"]] <- renderUI({
+        tags$table(
+            tags$thead(
+                tags$tr(
+                    lapply(
+                        c("File Name", "Sequence Length", "Protein", "State", ""),
+                        tags$td,
+                        style = "width: 20%;"
+                    )
+                )
+            ),
+            tags$tbody(
+                lapply(
+                    files_meta(),
+                    function(sfim) {
+                        # TODO: server generating function.
+                        input_summary_row_ui(sfim)
+                    }
+                )
+            )
+        )
+    })
     # TODO: import summary table generation from the files_meta (clear the meta after generating the table?).
     # TODO: delete button observers and state upadters generation.
 }